Optimisation du code gestion de connexion

+++ index.php
@@ -14,7 +14,6 @@ try{
 	{
 		if (strstr($_GET['action'], 'Admin'))
 		{
-			
 			if (!isset($_SESSION['id']) || !isset($_SESSION['pseudo']))
 			{	
 			throw new Exception ("Cette action est réservée aux administrateurs");
@@ -25,15 +24,13 @@ try{
 
 		switch($_GET['action']) {
 			
-			case 'nextPost':
-				getNextPost($_GET['id']);
-				break;
-			
+		/* PAGE ACCUEIL */
 			
 			case 'listPosts':
-
 				listPosts();
 				break;
+			
+		/* PAGE CHAPITRE */
 			
 			case 'viewPost':		
 				
@@ -45,6 +42,10 @@ try{
 				{
 					throw new Exception('Aucun identifiant de billet envoyé');
 				}
+				break;
+				
+				case 'nextPost':
+				getNextPost($_GET['id']);
 				break;
 				
 			case 'addComment':
@@ -79,24 +80,44 @@ try{
 				break;
 				
 				
+		/* GESTION CONNEXION */
+			
+			case 'authenticize':
+				if (!empty($_POST['identifiant']) && !empty($_POST['password'])){
+					authenticize($_POST['identifiant'], $_POST['password']);
+				}
+				else{
+					throw new Exception ('Login ou mot de passe incorrect');
+				}
+				break;				
+					
+			case 'login':
+				login();
+				break;
+			
+			case 'logoff':	
+				logoff();
+				break;	
 				
-				/* FONCTIONS ADMINISTRATEUR */
+				
+		/* - - - - - FONCTIONS ADMINISTRATEUR - - - - - - */
 					
-					
+		/* Vue d'ensemble */
+
+			case 'gestionAdmin':
+				listPostsAdmin();
+				break;		
+				
+			
+			
+		/* Modération commentaires */
+			
 			case 'moderationAdmin':
 				reportedComment();
 				break;
-				
-			case 'newPostAdmin':
-				require('view/admin/newPost.php');
-				break;
-			
-			case 'gestionAdmin':
-				listPostsAdmin();
-				break;
 			
 			case 'deleteCommentAdmin':
-				if ((isset($_GET['id']) ))
+				if (isset($_GET['id']) && $_GET['id'] > 0 )
 				{
 					deleteComment($_GET['id']);
 					reportedComment();
@@ -104,22 +125,29 @@ try{
 				else{
 					throw new Exception('Aucun identifiant de commentaire envoyé');
 				}
-			
-			break;
-			
+				break;
+				
+				
 			case 'deleteReportAdmin':
-					if (isset($_GET['id']) && $_GET['id'] > 0)
-					{
-						deleteReport($_GET['id']);
-						reportedComment();
-					}
-					else{
-						throw new Exception('Aucun identifiant de commentaire envoyé');
-					}
-					break;		
+				if (isset($_GET['id']) && $_GET['id'] > 0)
+				{
+					deleteReport($_GET['id']);
+					reportedComment();
+				}
+				else
+				{
+					throw new Exception('Aucun identifiant de commentaire envoyé');
+				}
+				break;	
+			
+		/* Gestion des billets */
+			
+			case 'newPostAdmin':
+				require('view/admin/newPost.php');
+				break;
 			
 			case 'deletePostAdmin':
-				if ((isset($_GET['id'])))
+				if (isset($_GET['id']) && $_GET['id'] > 0)
 					{
 						deletePost($_GET['id']);
 						deleteAllComments($_GET['id']);
@@ -130,32 +158,30 @@ try{
 				break;
 			
 			case 'updatePostAdmin':
-			if (isset($_GET['id']) && $_GET['id'] > 0)
-			{
-				openPost($_GET['id']);	
-			}
-			else
-			{
-				throw new Exception('Aucun identifiant de billet envoyé');
-			}
+				if (isset($_GET['id']) && $_GET['id'] > 0)
+				{
+					openPost($_GET['id']);	
+				}
+				else
+				{
+					throw new Exception('Aucun identifiant de billet envoyé');
+				}
 				break;
 			
 			case 'updatePostValidAdmin':
 				if(!empty($_POST['idPost']) && !empty($_POST['title']) && !empty($_POST['resume']) && !empty($_POST['content']))
 				{
-						updatePost($_POST['idPost'], $_POST['title'], $_POST['resume'], $_POST['content']);	
+					updatePost($_POST['idPost'], $_POST['title'], $_POST['resume'], $_POST['content']);	
 				}
-				else{
+				else
+				{
 					throw new Exception('Un paramètre est manquant');
 				}
 				break;				
-			
-			
-				
-				
+
 			case 'addBilletAdmin':
 			if(!empty($_POST['title']) && !empty($_POST['resume']) && !empty($_POST['content'])){
-				addBillet($_POST['title'], $_POST['resume'], $_POST['content']);
+				addPost($_POST['title'], $_POST['resume'], $_POST['content']);
 			}
 			else{
 				throw new Exception ('Un paramètre est manquant');
@@ -164,30 +190,7 @@ try{
 			
 			
 			
-			/* GESTION CONNEXION */
-			
-			case 'authenticize':
-				if (!empty($_POST['identifiant']) && !empty($_POST['password'])){
-					authenticize($_POST['identifiant'], $_POST['password']);
-				}
-				else{
-					throw new Exception ('Login ou mot de passe incorrect');
-				}
-				break;				
-			
-			
-			case 'login':
-				require('view/login.php');
-				break;
-			
-			
-			case 'logoff':	
-				logoff();
-				break;	
-
-				
-				
-
+		/* Gestion de page inconnue */	
 			default:
 				throw new Exception('Oups ! La page demandée n\'existe pas.');
 				break;
@@ -195,13 +198,13 @@ try{
 		}
 		
 	}
-	
+		/* Page d'index par défaut */
 	else
 	{
 		listPosts();
 	}
 }
-
+		/* Affichage de la page d'erreur */
 catch(Exception $e){
 	  $errorMessage = $e->getMessage();
 	  require("view/erreur.php");
